requirements.txt
pillow
pyTelegramBotAPI